Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

KnetNLPModels.KnetNLPModel

KnetNLPModels.accuracy

KnetNLPModels.build_array

KnetNLPModels.build_nested_array_from_vec

KnetNLPModels.build_nested_array_from_vec

KnetNLPModels.create_minibatch

KnetNLPModels.reset_minibatch_test!

KnetNLPModels.reset_minibatch_train!

KnetNLPModels.set_vars!

KnetNLPModels.vcat_arrays_vector

KnetNLPModels.vector_params

NLPModels.grad!

NLPModels.obj

​

KnetNLPModels.KnetNLPModel — Type

KnetNLPModel

Data structure that interfaces neural networks defined with Knet.jl as an NLPModel.

KnetNLPModels.accuracy — Method

accuracy(nlp)

Computes the accuracy of the network nlp.chain given the data in nlp.minibatch_test.

KnetNLPModels.build_array — Method

build_array(vec, var_layer, index)

Inverse of the function Knet.cat1d, it generates a KnetArray similar to varlayer from the vector vec. The values are those of the vector in the range of index to index+consumedindex. This method is not optimised, it consumes memory.

KnetNLPModels.build_nested_array_from_vec — Method

build_nested_array_from_vec(model, v)

Build a vector of KnetArrays from v similar to Knet.params(model.chain). Call build_array iteratively to build each intermediary KnetArrays. This methods is not optimised, it consumes memory.

KnetNLPModels.build_nested_array_from_vec — Method

build_nested_array_from_vec(chain, v)

Builds a vector of KnetArrays of the same shape as chain of value v. It calls iteratively build_array to build each intermediary KnetArrays. This method is not optimised, it consumes memory.

KnetNLPModels.create_minibatch — Method

create_minibatch(X, Y, minibatch_size)

Create a minibatch of the data (X,Y) of size minibatch_size

KnetNLPModels.reset_minibatch_test! — Method

reset_minibatch_test!(nlp)

Take a new minibatch for the KnetNLPModel. Usually use before a new accuracy test.

KnetNLPModels.reset_minibatch_train! — Method

reset_minibatch_train!(nlp)

Take a new minibatch for the KnetNLPModel. Usually use before a new evaluation.

KnetNLPModels.set_vars! — Method

set_vars!(model, new_w)

set_vars!(chain, new_w)

Set the variables of model or chain to new_w. Build a vector of KnetArrays from v similar to Knet.params(model.chain). Then it sets these variables to the nested array.

KnetNLPModels.vcat_arrays_vector — Method

vcat_arrays_vector(arrays_vector)

Flatten a vector of arrays to a vector. It keeps the order induce by Knet.cat1d to flatten an array.

KnetNLPModels.vector_params — Method

vector_params(model)

Retrieves the variables as a vector from the vector of variable's arrays that composes model.

NLPModels.grad! — Method

g = grad!(nlp, x, g)

Evaluate $∇f(x)$, the gradient of the objective function at x in place.

NLPModels.obj — Method

f = obj(nlp, x)

Evaluate $f(x)$, the objective function of nlp at x.