Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

KnetNLPModels.KnetNLPModel

KnetNLPModels.KnetNLPModel

KnetNLPModels.accuracy

KnetNLPModels.build_layer_from_vec

KnetNLPModels.build_layer_from_vec!

KnetNLPModels.build_nested_array_from_vec

KnetNLPModels.build_nested_array_from_vec!

KnetNLPModels.create_minibatch

KnetNLPModels.reset_minibatch_test!

KnetNLPModels.reset_minibatch_train!

KnetNLPModels.set_size_minibatch!

KnetNLPModels.set_vars!

KnetNLPModels.vcat_arrays_vector

KnetNLPModels.vector_params

NLPModels.grad!

NLPModels.obj

​

KnetNLPModels.KnetNLPModel — Type

KnetNLPModel{T, S, C <: Chain} <: AbstractNLPModel{T, S}

Data structure that makes the interfaces between neural networks defined with Knet.jl and NLPModels.

KnetNLPModels.KnetNLPModel — Method

  KnetNLPModel(chain_ANN, size_minibatch; data_train=data_train, data_test=data_test)

Build a KnetNLPModel from the neural network represented by chain_ANN. chain is build by Knet.jl, see the tutorial for more details. The other mandatory data are: data_train, data_test and the size of the minibatch size_minibatch. By default they are set to MNIST dataset with minibatchs of size 100.

KnetNLPModels.accuracy — Method

accuracy(nlp :: KnetNLPModel{T, S, C}) where {T, S, C}

Computes the accuracy of the network nlp.chain given the data in nlp.minibatch_test. The accuracy is based from the whole testing dataset.

KnetNLPModels.build_layer_from_vec! — Method

build_layer_from_vec!(cuArray :: CuArray{T, N, CUDA.Mem.DeviceBuffer} where N, v :: Vector{T}, index :: Int) where {T <: Number}

Inverse of the function Knet.cat1d, it sets cuArray to the values of vec in the range index+1:index+consumed_index.

KnetNLPModels.build_layer_from_vec — Method

build_layer_from_vec(v :: Vector{T}, var_layers :: CuArray{T, N, CUDA.Mem.DeviceBuffer} where N, index :: Int) where {T <: Number}

Inverse of the function Knet.cat1d, it builds a CuArray similar to var_layers from vec. The return values are those of the vector in the range of index+1:index+consumed_index. This method is not optimised, it consumes memory.

KnetNLPModels.build_nested_array_from_vec! — Method

build_nested_array_from_vec!(vec_CuArray :: Vector{CuArray{T, N, CUDA.Mem.DeviceBuffer} where N}, v :: Vector{T}) where {T <: Number}

	build_nested_array_from_vec!(model :: KnetNLPModel{T, S, C}, new_w :: Vector) where {T, S, C}

Build a vector of CuArrays from v similar to Knet.params(model.chain) or Knet.params(chain_ANN). It calls iteratively build_layer_from_vec to build each intermediary CuArray. This method is not optimised, it consumes memory.

KnetNLPModels.build_nested_array_from_vec — Method

build_nested_array_from_vec(chain_ANN :: C, v :: Vector{T}) where {C <: Chain, T <: Number}

	build_nested_array_from_vec(model :: KnetNLPModel{T, S, C}, v :: Vector{T}) where {T, S, C}

Build a vector of KnetArrays from v similar to Knet.params(model.chain) or 'Knet.params(chainANN)'. It calls iteratively `buildlayerfromvecto build each intermediaryCuArrays`. This method is not optimised, it consumes memory.

KnetNLPModels.create_minibatch — Method

create_minibatch(X, Y, minibatch_size)

Create a minibatch of the data X, Y of size minibatch_size.

KnetNLPModels.reset_minibatch_test! — Method

reset_minibatch_test!(nlp :: KnetNLPModel{T, S, C}) where {T, S, C}

Take a new testing minibatch for the KnetNLPModel. Usually use before a new accuracy test.

KnetNLPModels.reset_minibatch_train! — Method

reset_minibatch_train!(nlp :: KnetNLPModel{T, S, C}) where {T, S, C}

Take a new training minibatch for the KnetNLPModel. Usually use before a new evaluation.

KnetNLPModels.set_size_minibatch! — Method

set_size_minibatch!(knetnlp, size_minibatch)

Change the size of the minibatchs of training and testing of the knetnlp. After a call of set_size_minibatch!, if one want to use a minibatch of size size_minibatch it must use beforehand reset_minibatch_train!.

KnetNLPModels.set_vars! — Method

set_vars!(model :: KnetNLPModel{T, S, C}, new_w :: Vector) where {T, S, C}

set_vars!(chain_ANN :: C, nested_w :: Vector{CuArray{T, N, CUDA.Mem.DeviceBuffer} where N}) where {C <: Chain, T <: Number}

Set the variables of model or chain to new_w. Build a vector of CuArrays from v similar to Knet.params(model.chain). Then it sets these variables to the nested array.

KnetNLPModels.vcat_arrays_vector — Method

vcat_arrays_vector(arrays_vector)

Flatten a vector of arrays to a vector. It concatenate the variables produce by applying Knet.cat1d to each array.

KnetNLPModels.vector_params — Method

vector_params(chain :: C) where C <: Chain

	vector_params(nlp :: KnetNLPModel)

Retrieves the variables within chain or nlp.chain as a vector.

NLPModels.grad! — Method

g = grad!(nlp, x, g)

Evaluate ∇f(x), the gradient of the objective function at x in place.

NLPModels.obj — Method

f = obj(nlp, x)

Evaluate f(x), the objective function of nlp at x.