Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

KnetNLPModels.KnetNLPModel

KnetNLPModels.KnetNLPModel

KnetNLPModels.accuracy

KnetNLPModels.build_layer_from_vec!

KnetNLPModels.build_nested_array_from_vec

KnetNLPModels.build_nested_array_from_vec!

KnetNLPModels.create_minibatch

KnetNLPModels.flag_dim

KnetNLPModels.reset_minibatch_test!

KnetNLPModels.reset_minibatch_train!

KnetNLPModels.set_size_minibatch!

KnetNLPModels.set_vars!

KnetNLPModels.vcat_arrays_vector

KnetNLPModels.vector_params

NLPModels.grad!

NLPModels.obj

​

KnetNLPModels.KnetNLPModel — Type

KnetNLPModel{T, S, C <: Chain} <: AbstractNLPModel{T, S}

Data structure that makes the interfaces between neural networks defined with Knet.jl and NLPModels. A KnetNLPModel has fields

meta and counters retain informations about the KnetNLPModel;

chain is the chained structure representing the neural network;

data_train is the complete data training set;

data_test is the complete data test;

size_minibatch parametrizes the size of an training and test minibatches, which are of size 1/size_minibatch * length(ytrn) and 1/size_minibatch * length(ytst);

training_minibatch_iterator is an iterator over an training minibatches;

test_minibatch_iterator is an iterator over the test minibatches;

current_training_minibatch is the training minibatch used to evaluate the neural network;

current_minibatch_test is the current test minibatch, it is not used in practice;

w is the vector of weights/variables;

layers_g is a nested array used for internal purposes;

nested_array is a vector of Array{T,N}; its shape matches that of chain.

KnetNLPModels.KnetNLPModel — Method

KnetNLPModel(chain_ANN; size_minibatch=100, data_train=MLDatasets.MNIST.traindata(Float32), data_test=MLDatasets.MNIST.testdata(Float32))

Build a KnetNLPModel from the neural network represented by chain_ANN. chain_ANN is built using Knet.jl, see the tutorial for more details. The other data required are: an iterator over the training dataset data_train, an iterator over the test dataset data_test and the size of the minibatch size_minibatch. Suppose (xtrn,ytrn) = knetnlp.data_train, then the size of each training minibatch will be 1/size_minibatch * length(ytrn). By default, the other data are respectively set to the training dataset and test dataset of MLDatasets.MNIST, with each minibatch a hundredth of the dataset.

KnetNLPModels.accuracy — Method

accuracy(nlp::AbstractKnetNLPModel)

Compute the accuracy of the network nlp.chain on the entire test dataset.

KnetNLPModels.build_layer_from_vec! — Method

build_layer_from_vec!(array :: AbstractArray{T, N}, v :: AbstractVector{T}, index :: Int) where {T <: Number, N}

Inverse of the function Knet.cat1d; set array to the values of v in the range index+1:index+consumed_index.

KnetNLPModels.build_nested_array_from_vec! — Method

build_nested_array_from_vec!(model::AbstractKnetNLPModel{T,S}, new_w::AbstractVector{T}) where {T, S}
build_nested_array_from_vec!(nested_array :: AbstractVector{<:AbstractArray{T,N} where {N}}, new_w :: AbstractVector{T}) where {T <: Number}

Build a vector of AbstractArray from new_w similar to Knet.params(model.chain) or nested_array. Call iteratively build_layer_from_vec! to build each intermediate AbstractArray. This method is not optimized; it allocates memory.

KnetNLPModels.build_nested_array_from_vec — Method

nested_array = build_nested_array_from_vec(model::AbstractKnetNLPModel{T, S}, v::AbstractVector{T}) where {T <: Number, S}
nested_array = build_nested_array_from_vec(chain_ANN::C, v::AbstractVector{T}) where {C <: Chain, T <: Number}
nested_array = build_nested_array_from_vec(nested_array::AbstractVector{<:AbstractArray{T,N} where {N}}, v::AbstractVector{T}) where {T <: Number}

Build a vector of AbstractArray from v similar to Knet.params(model.chain), Knet.params(chain_ANN) or nested_array. Call iteratively build_layer_from_vec to build each intermediate AbstractArray. This method is not optimized; it allocates memory.

KnetNLPModels.create_minibatch — Method

create_minibatch(X, Y, minibatch_size)

Create a minibatch's iterator of the data X, Y of size 1/minibatch_size * length(Y).

KnetNLPModels.flag_dim — Method

flag_dim(x)

Returns true if x has 3 dimensions. This function is used to reshape X in create_minibatch(X, Y, minibatch_size) in case x has only 3 dimensions.

KnetNLPModels.reset_minibatch_test! — Method

reset_minibatch_test!(nlp::AbstractKnetNLPModel)

Select a new test minibatch for nlp.

KnetNLPModels.reset_minibatch_train! — Method

reset_minibatch_train!(nlp::AbstractKnetNLPModel)

Select a new training minibatch for nlp. Typically used before a new evaluation of the loss function/gradient.

KnetNLPModels.set_size_minibatch! — Method

set_size_minibatch!(knetnlp::AbstractKnetNLPModel, size_minibatch::Int)

Change the size of both training and test minibatches of the knetnlp. Suppose (xtrn,ytrn) = knetnlp.data_train, then the size of each training minibatch will be 1/size_minibatch * length(ytrn); the test minibatch follows the same logic. After a call of set_size_minibatch!, you must call reset_minibatch_train!(knetnlp) to use a minibatch of the expected size.

KnetNLPModels.set_vars! — Method

set_vars!(model::AbstractKnetNLPModel{T,S}, new_w::AbstractVector{T}) where {T<:Number, S}
set_vars!(chain_ANN :: C, nested_w :: AbstractVector{<:AbstractArray{T,N} where {N}}) where {C <: Chain, T <: Number}
set_vars!(vars :: Vector{Param}, nested_w :: AbstractVector{<:AbstractArray{T,N} where {N}})

)

Set the variables of model (resp. chain_ANN and vars) to new_w (resp. nested_w). Build nested_w: a vector of AbstractArray from new_v similar to Knet.params(model.chain). Then, set the variables vars of the neural netword model (resp. chain_ANN) to new_w (resp. nested_w). set_vars!(model, new_w) allocates memory.

KnetNLPModels.vcat_arrays_vector — Method

vcat_arrays_vector(arrays_vector::AbstractVector{Param})

Flatten a vector of arrays arrays_vector to a vector. It concatenates the vectors produced by the application of Knet.cat1d to each array.

KnetNLPModels.vector_params — Method

vector_params(chain :: C) where C <: Chain
vector_params(nlp :: AbstractKnetNLPModel)

Retrieve the variables within chain or nlp.chain as a vector.

NLPModels.grad! — Method

g = grad!(nlp, x, g)

Evaluate ∇f(x), the gradient of the objective function at x in place.

NLPModels.obj — Method

f = obj(nlp, x)

Evaluate f(x), the objective function of nlp at x.